% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{coef.gradient}
\alias{coef.gradient}
\title{Extract Model Coefficients}
\usage{
\method{coef}{gradient}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments passed to or from other methods}
}
\value{
list with coefficients
}
\description{
generic function which extracts model coefficients from objects returned by modeling functions
}
