% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{gd}
\alias{gd}
\title{Il wrap fatto a mano}
\usage{
gd(b, formula, stepsize = 0.001, maxit = 1000, tolerance = 0.001, verbose = T)
}
\arguments{
\item{b}{Beta init}

\item{formula}{formula}

\item{stepsize}{stepsize}

\item{maxit}{max iteration}

\item{tolerance}{tolerance}

\item{verbose}{verbose}
}
\description{
Il wrap fatto a mano
}
