% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_gradient_cv.R
\name{lm_gradient_cv}
\alias{lm_gradient_cv}
\title{Linear model gradient cross validation}
\usage{
lm_gradient_cv(
  n.folds,
  b,
  formula,
  data = NULL,
  maxit = 1000,
  tolerance = 1e-05,
  stepsize = 1e-04,
  fun = "sd",
  parallel = FALSE
)
}
\arguments{
\item{n.folds}{nfold}

\item{b}{b}

\item{formula}{formula}

\item{data}{data}

\item{maxit}{max ite}

\item{tolerance}{tol}

\item{stepsize}{step size}

\item{fun}{fun}
}
\value{
something
}
\description{
Linear model gradient cross validation
}
