% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_gradient_looc.R
\name{lm_gradient_looc}
\alias{lm_gradient_looc}
\title{Title}
\usage{
lm_gradient_looc(
  b,
  formula,
  data = NULL,
  maxit = 1000,
  tolerance = 1e-05,
  stepsize = 1e-04,
  fun = "sd",
  parallel = FALSE
)
}
\arguments{
\item{b}{b}

\item{formula}{formula}

\item{data}{data}

\item{maxit}{maxit}

\item{tolerance}{tol}

\item{stepsize}{a}

\item{fun}{a}

\item{parallel}{a}
}
\value{
a
}
\description{
Title
}
