% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{plot.gradient}
\alias{plot.gradient}
\title{Title}
\usage{
\method{plot}{gradient}(object, ...)
}
\arguments{
\item{object}{object}
\method{plot}{gradient}(x, ...)
}
\arguments{
\item{x}{obj}

\item{...}{others}
}
\value{a
}
\description{
Title
}
