% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{predict.gradient}
\alias{predict.gradient}
\title{Title}
\usage{
\method{predict}{gradient}(object, newdata, ...)
}
\arguments{
\item{object}{object}

\item{...}{others}
}
\value{

}
\description{
Title
}
