% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{print.gradient}
\alias{print.gradient}
\title{Print
prints its argument}
\usage{
\method{print}{gradient}(x, ...)
}
\arguments{
\item{x}{object of gradient class}

\item{...}{further arguments passed to or from other methods}
}
\value{
object of class gradient
}
\description{
Print
prints its argument
}
