% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd}
\alias{sd}
\title{Il wrap fatto a mano (SD però)}
\usage{
sd(b, formula, maxit = 1000, tolerance = 0.001, verbose = T)
}
\arguments{
\item{b}{Beta init}

\item{formula}{formula}

\item{maxit}{max iteration}

\item{tolerance}{tole}

\item{verbose}{verbose}
}
\description{
Il wrap fatto a mano (SD però)
}
