% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{summary.gradient}
\alias{summary.gradient}
\title{Summary}
\usage{
\method{summary}{gradient}(object, ...)
}
\arguments{
\item{object}{object of descent class}

\item{...}{further arguments passed to or from other methods}
}
\value{
object of class descent
}
\description{
Summarize an analysis of the fitted model
}
