% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_cv_methods.R
\name{summary.gradient_cv}
\alias{summary.gradient_cv}
\title{Title}
\usage{
\method{summary}{gradient_cv}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{others}
}
\value{
a
}
\description{
Title
}
