% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_optim_SQGDE.R
\name{update_optim_SQGDE}
\alias{update_optim_SQGDE}
\title{update_optim_SQGDE}
\usage{
update_optim_SQGDE(
  ObjFun,
  last_old_particles,
  last_old_weights,
  control_params = GetAlgoParams(),
  ...
)
}
\arguments{
\item{ObjFun}{The same ObjFun used in optim_SQGDE. Must have the same length as the former run.}

\item{last_old_particles}{The last valid iteration of a particle trace container from the output of optim_SQGDE.
Requires the same dimensions of the particles and ObjFun as former run.}

\item{last_old_weights}{The last valid iteration of a weights trace container from the output of optim_SQGDE.
Requires the same dimensions of the particles and parameters as former run.}

\item{control_params}{A list of control parameters for the SQG-DE algorithm as used in optim_SQGDE. The primary required
parameter is \code{n_params}. For more details, see the documentation of \code{\link{GetAlgoParams}}.}

\item{...}{Additional arguments to be passed to \code{ObjFun}.}
}
\value{
A list containing:
\itemize{
\item \code{solution}: The best solution found by the algorithm.
\item \code{weight}: The corresponding weight (i.e., \code{f(solution)}).
\item \code{particles_trace} (optional): Trace of particle positions over iterations (if \code{return_trace = TRUE}).
\item \code{weights_trace} (optional): Trace of weights over iterations (if \code{return_trace = TRUE}).
\item \code{converged}: Logical, indicating whether the convergence criterion was met.
}
}
\description{
Updates a set of particle traces and weights for further optimization.
Useful when convergence criteria is not met in the alotted iterations or when
only particle save file is recovered.
Uses Stochastic Quasi-Gradient Differential Evolution (SQG-DE;
Sala, Baldanzini, and Pierini, 2018) to minimize a specified objective function \code{f(x)}.
To maximize a function, pass \code{g(x) = -f(x)} as the \code{ObjFun} argument.
}
\examples{
##############
# Maximum Likelihood Example
##############

# Simulate data
dataExample <- matrix(rnorm(1000, c(-1, 1, 0, 1), c(1, 1, 1, 1)), ncol = 4, byrow = TRUE)

# Define parameter names
param_names_example <- c("mu_1", "mu_2", "mu_3", "mu_4")

# Define negative log-likelihood function
ExampleObjFun <- function(x, data, param_names) {
  names(x) <- param_names
  -sum(dnorm(data[, 1], x["mu_1"], sd = 1, log = TRUE) +
       dnorm(data[, 2], x["mu_2"], sd = 1, log = TRUE) +
       dnorm(data[, 3], x["mu_3"], sd = 1, log = TRUE) +
       dnorm(data[, 4], x["mu_4"], sd = 1, log = TRUE))
}

# Run optimization
out <- optim_SQGDE(
  ObjFun = ExampleObjFun,
  control_params = GetAlgoParams(
    n_params = length(param_names_example),
    n_iter = 250,
    n_particles = 12,
    n_diff = 2,
    return_trace = TRUE
  ),
  data = dataExample,
  param_names = param_names_example
)

# Plot particle trajectories
par(mfrow = c(2, 2))
matplot(out$particles_trace[, , 1], type = 'l')
matplot(out$particles_trace[, , 2], type = 'l')
matplot(out$particles_trace[, , 3], type = 'l')
matplot(out$particles_trace[, , 4], type = 'l')

# SQG-DE solution
out$solution

# Analytical solution
apply(dataExample, 2, mean)
}
