% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{validate.gradient}
\alias{validate.gradient}
\title{Title}
\usage{
validate.gradient(object, newdata, y, method = "all", ...)
}
\arguments{
\item{object}{object}

\item{...}{others}
}
\value{

}
\description{
Title
}
