% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_gradient.R
\name{lm_gradient}
\alias{lm_gradient}
\title{Title}
\usage{
lm_gradient(
  b,
  formula,
  data = NULL,
  maxit = 1000,
  tolerance = 1e-05,
  stepsize = 1e-04,
  fun = "sd",
  verbose = T
)
}
\arguments{
\item{b}{beta}

\item{formula}{formula}

\item{data}{data}

\item{maxit}{max iterations}

\item{tolerance}{tolerance}

\item{stepsize}{step size}

\item{fun}{function}

\item{verbose}{verbose}
}
\value{
results
}
\description{
Title
}
