% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_gradient_looc.R
\name{lm_gradient_looc}
\alias{lm_gradient_looc}
\title{Leave one out cross validation for Linear Models with gradient}
\usage{
lm_gradient_looc(
  b,
  formula,
  data = NULL,
  maxit = 1000,
  tolerance = 1e-05,
  stepsize = 1e-04,
  fun = "sd",
  parallel = FALSE
)
}
\arguments{
\item{b}{initial values for beta}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted}

\item{data}{an optional data frame, list or environment. If not found in data, the variables are taken from environment(formula)}

\item{maxit}{number of max iterations of the algorithm}

\item{tolerance}{the algorithm will stop if tolerance is reached}

\item{stepsize}{size of each step (only for gradient descent)}

\item{fun}{function can be \code{sd} for steepest descent or \code{gd} for gradient descent}

\item{parallel}{perform the cross validation in multi-core}
}
\value{
the function returns an object of class \code{"gradient"}
}
\description{
\code{lm_gradient_looc} is used to perform a leave one out cross validation to a linear model with gradient or steepest descent.
}
