% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_SQGDE.R
\name{optim_SQGDE}
\alias{optim_SQGDE}
\title{optim_SQGDE}
\usage{
optim_SQGDE(ObjFun, control_params = GetAlgoParams(), ...)
}
\arguments{
\item{ObjFun}{A scalar-returning function or a list of functions to minimize. Each function
should accept a real-valued vector of length \code{n_params} as input. When providing a list of functions,
block updating is enabled, iterating through each function to update only a subset of parameters conditional (see \code{param_ind_to_update_list} in \code{\link{GetAlgoParams}}).}

\item{control_params}{A list of control parameters for the SQG-DE algorithm. The primary required
parameter is \code{n_params}. For more details, see the documentation of \code{\link{GetAlgoParams}}.}

\item{...}{Additional arguments to be passed to \code{ObjFun}.}
}
\value{
A list containing:
\itemize{
\item \code{solution}: The best solution found by the algorithm.
\item \code{weight}: The corresponding weight (i.e., \code{f(solution)}).
\item \code{particles_trace} (optional): Trace of particle positions over iterations (if \code{return_trace = TRUE}).
\item \code{weights_trace} (optional): Trace of weights over iterations (if \code{return_trace = TRUE}).
\item \code{converged}: Logical, indicating whether the convergence criterion was met.
}
}
\description{
Uses Stochastic Quasi-Gradient Differential Evolution (SQG-DE;
Sala, Baldanzini, and Pierini, 2018) to minimize a specified objective function \code{f(x)}.
To maximize a function, pass \code{g(x) = -f(x)} as the \code{ObjFun} argument.
}
\examples{
##############
# Maximum Likelihood Example
##############

# Simulate data
dataExample <- matrix(rnorm(1000, c(-1, 1, 0, 1), c(1, 1, 1, 1)), ncol = 4, byrow = TRUE)

# Define parameter names
param_names_example <- c("mu_1", "mu_2", "mu_3", "mu_4")

# Define negative log-likelihood function
ExampleObjFun <- function(x, data, param_names) {
  names(x) <- param_names
  -sum(dnorm(data[, 1], x["mu_1"], sd = 1, log = TRUE) +
       dnorm(data[, 2], x["mu_2"], sd = 1, log = TRUE) +
       dnorm(data[, 3], x["mu_3"], sd = 1, log = TRUE) +
       dnorm(data[, 4], x["mu_4"], sd = 1, log = TRUE))
}

# Run optimization
out <- optim_SQGDE(
  ObjFun = ExampleObjFun,
  control_params = GetAlgoParams(
    n_params = length(param_names_example),
    n_iter = 250,
    n_particles = 12,
    n_diff = 2,
    return_trace = TRUE
  ),
  data = dataExample,
  param_names = param_names_example
)

# Plot particle trajectories
par(mfrow = c(2, 2))
matplot(out$particles_trace[, , 1], type = 'l')
matplot(out$particles_trace[, , 2], type = 'l')
matplot(out$particles_trace[, , 3], type = 'l')
matplot(out$particles_trace[, , 4], type = 'l')

# SQG-DE solution
out$solution

# Analytical solution
apply(dataExample, 2, mean)
}
