% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{plot.gradient}
\alias{plot.gradient}
\title{Plot descent convergence}
\usage{
\method{plot}{gradient}(x, ...)
}
\arguments{
\item{x}{object of descent class}

\item{...}{further arguments passed to or from other methods}
}
\value{
}
\description{
Plot descent convergence
}
