% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{predict.gradient}
\alias{predict.gradient}
\title{Title}
\usage{
\method{predict}{gradient}(object, newdata, ...)
}
\arguments{
\item{object}{object}

<<<<<<< HEAD
\item{...}{others}
}
\value{

=======
\item{newdata}{new data}

\item{...}{others}
}
\value{
a
>>>>>>> ae40a4bb2e6266cc29f7955e4dc8adda667bf855
}
\description{
Title
}
