% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_methods.R
\name{summary.gradient}
\alias{summary.gradient}
\title{Title}
\usage{
\method{summary}{gradient}(object, ...)
}
\arguments{
\item{object}{oggetto}

\item{...}{others}
}
\value{

}
\description{
Title
}
