% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_cv_methods.R
\name{summary.gradient_cv}
\alias{summary.gradient_cv}
\title{Summary}
\usage{
\method{summary}{gradient_cv}(object, ...)
}
\arguments{
\item{object}{object of gradient_cv class}

\item{...}{further arguments passed to or from other methods}
}
\value{
object of class gradient_cv
}
\description{
Summarize an analysis of the fitted model
}
